pyproject.toml
[project]
name = "wikibots"
version = "1.0.0"
authors = [
    {name = "DaxServer",email = "daxserver@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "deepdiff (>=9.0.0,<10.0.0)",
    "requests-oauthlib (>=2.0.0,<3.0.0)",
    "google-api-python-client (>=2.166.0,<3.0.0)",
    "flickr-photos-api (>=3.0.0,<4.0.0)",
    "flickr-url-parser (>=1.11.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "mwparserfromhell (>=0.7.2,<1.0.0)",
    "requests (>=2.32.0,<3.0.0)",
]

[tool.poetry]
packages = [{include = "wikibots", from = "src"}]

[project.scripts]
flickr = "wikibots.flickr:main"
inaturalist = "wikibots.inaturalist:main"
pas = "wikibots.pas:main"
youtube = "wikibots.youtube:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[io.buildpacks]
exclude = [
    ".poetry",
    ".venv",
    "apicache",
    "*.sh",
    "user-password.py",
    "throttle"
]

[dependency-groups]
dev = [
    "ruff (>=0.15.9,<0.16.0)",
    "ty (>=0.0.28,<0.0.29)",
    "isort (>=8.0.1,<9.0.0)"
]

[tool.isort]
profile = "black"
